// web/src/api/generated/model/documentKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type DocumentKind = typeof DocumentKind[keyof typeof DocumentKind];


export const DocumentKind = {
  Quote: 'Quote',
  SalesOrder: 'SalesOrder',
  DeliveryNote: 'DeliveryNote',
  SalesInvoice: 'SalesInvoice',
  SalesCreditNote: 'SalesCreditNote',
  PurchaseOrder: 'PurchaseOrder',
  GoodsReceipt: 'GoodsReceipt',
  PurchaseInvoice: 'PurchaseInvoice',
  PurchaseDebitNote: 'PurchaseDebitNote',
} as const;