// web/src/api/generated/model/documentStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  Draft: 'Draft',
  Issued: 'Issued',
  Converted: 'Converted',
} as const;