// web/src/api/generated/model/documentSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { DocumentKind } from './documentKind';
import type { DocumentStatus } from './documentStatus';

export interface DocumentSummary {
  id: string;
  kind: DocumentKind;
  /** @nullable */
  number: string | null;
  date: string;
  /** @nullable */
  dueDate: string | null;
  status: DocumentStatus;
  partyId: string;
  currencyCode: string;
  total: number;
  /** @nullable */
  reference: string | null;
}